requirements.txt
streamlit
python-dotenv
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface
moviepy 
pydub 
SpeechRecognition 
requests
openpyxl
